26/9/24 today summit report of last 3 month of admin panel , created pull request for partner signin & signup page , table revamp of user section 2 table(studio booking, service booking)
